requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
requests>=2.31.0
pika>=1.3.0
python-multipart>=0.0.6
httpx>=0.25.0